fix: preparing for like and comment

+++ BackendServer/src/community/posts/posts.controller.ts
@@ -1,26 +1,57 @@
 // src/posts/posts.controller.ts
-import { Controller, Get, Post, Body, Query, Param, ParseIntPipe, Delete, UseGuards, Req } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Query,
+  Param,
+  ParseIntPipe,
+  Delete,
+  UseGuards,
+  Req,
+  Logger,
+  ForbiddenException,
+} from '@nestjs/common';
 import { PostsService } from './posts.service';
 import { CreatePostDto } from './dto/create-post.dto';
 import { Post as PostEntity } from '../entities/post.entity';
 import { AuthGuard } from '@nestjs/passport';
+import { Request } from 'express';
 
 @Controller('posts')
 export class PostsController {
   constructor(private readonly postsService: PostsService) {}
+  private readonly logger = new Logger(PostsController.name);
 
   @Post()
   @UseGuards(AuthGuard('jwt'))
-  create(@Body() createPostDto: CreatePostDto, @Req() req): Promise<PostEntity> {
+  create(
+    @Body() createPostDto: CreatePostDto,
+    @Req() req: Request,
+  ): Promise<PostEntity> {
+    if (!req.user || !req.user.id) {
+      this.logger.error(
+        'Authentication information missing from request user object.',
+      );
+      throw new ForbiddenException('사용자 인증 정보가 없습니다.');
+    }
     // ⭐️ req.user 객체 전체를 서비스로 전달합니다.
     return this.postsService.create(createPostDto, req.user);
   }
 
   @Get()
-  @UseGuards(AuthGuard('jwt')) // 비로그인 사용자도 허용하는 가드
-  findAll(@Query('category') category: string, @Req() req) {
+  @UseGuards(AuthGuard('jwt'))
+  findAll(@Query('category') category: string, @Req() req: Request) {
     // ⭐️ req.user 객체 전체를 서비스로 전달합니다.
-    return this.postsService.findAll(category, req.user);
+    if (!req.user || !req.user.id) {
+      this.logger.error(
+        'Authentication information missing from request user object.',
+      );
+      throw new ForbiddenException('사용자 인증 정보가 없습니다.');
+    }
+
+    return this.postsService.findAllPostsWithLikeStatus(category, req.user);
   }
 
   @Get(':id')
@@ -30,15 +61,14 @@ export class PostsController {
 
   @Delete(':id')
   @UseGuards(AuthGuard('jwt'))
-  remove(@Param('id', ParseIntPipe) id: number, @Req() req) {
+  remove(@Param('id', ParseIntPipe) id: number, @Req() req: Request) {
+    if (!req.user || !req.user.id) {
+      this.logger.error(
+        'Authentication information missing from request user object.',
+      );
+      throw new ForbiddenException('사용자 인증 정보가 없습니다.');
+    }
     // ⭐️ req.user 객체 전체를 서비스로 전달합니다.
     return this.postsService.remove(id, req.user);
-  }
-
-  @Post(':id/like')
-  @UseGuards(AuthGuard('jwt'))
-  toggleLike(@Param('id', ParseIntPipe) id: number, @Req() req) {
-    // ⭐️ req.user 객체 전체를 서비스로 전달합니다.
-    return this.postsService.toggleLike(id, req.user);
   }
 }
